pyproject.toml
[project]
name = "diary"
version = "0.1.0"
description = ""
authors = [
    {name = "sorul",email = "cromerovargas2d@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client (>=2.176.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "folium (>=0.20.0,<0.21.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
